Improve separation of cuda code from cpu-only code
This comes at the cost of one additional indirection (virtual function call)
for each call to 'CudaPathTracer::render', but I think this must happen
eventually in order to separate the two worlds.
Benefits are less if-defs and also the possibility to better layout the cuda
classes design. E.g., I think the next step would be to merge
CudaPathTracerImpl with CudaRenderer.

To be reviewed by @mkost