Make repo identical to original after histroy rewrite
The history needed to be rewritten to get rid of the
140mb cube.bmp file which wouldn't be allowed by
github to be part of the repo. Turns out, history
rewriting is not that simple in git. However, I think
I got it right now except for the 3 minor mishaps
fixed by this commit. If I diff the directory
contents of the worktree with the final version
available to me (dev branch) to this commit's repo
state I get the following:

Only in pray: build
Only in pray-dev/examples/cube: cube.bmp
File pray/.git is a directory while file pray-dev/.git is a regular file

I.e. the cube.bmp is gone (yay!), the build dir is
seen by diff but not checked into git and pray-dev
is a worktree instead of a full repo clone. Perfect!